Update README.md
This pull request comes with only cosmetic changes to the text of the main README. I will have to check, however, if all the doctests pass, as I am afraid this does not have to be the case. I am also pondering whether all the examples, particularlt for `map_object()`, are simple enough. I will take care of that in a next branch.